+++ CHANGES.rst
@@ -15,6 +15,8 @@ Migrations
 Changes
 ~~~~~~~
 
+- Add a new CellCountryProvider to allow country searches based on cell data.
+
 - #406: Allow access to the country API via empty GET requests.
 
 - Use custom JSON encoding for kombu transport to support date/time objects.